Inject methods as files are loaded.
This allows a lite loading approach (e.g. you may be interested only on mainstream browsers, or you want to detect only bots).

By default, this is what is being loaded:

- Main browsers (Safari, Chrome, IE/Edge, Firefox, Opera)
- Meta info (only for what's currently available)

To load lite mode, you have to require `browser/lite.rb` instead. Then, you can load other pieces of Browser you need (e.g. bot, platform, device).
