+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Remove Browser::Bot.detect_empty_ua! in favor of Browser::Bot.allow_empty_ua!
 - Detect unknown platform/device as bots.
 - `browser.modern?` and `Browser.modern_rules` were removed; create a class that wraps the logic you want (e.g. a Rails helper method instantiates a class and passes the browser object to it.
+- Add lite mode, which will load only a few pieces of Browser, so you can further customize what you want to be loaded.
 
 ## v2.5.3
