+++ lib/browser/alipay.rb
@@ -1,6 +1,12 @@
 # frozen_string_literal: true
 
 module Browser
+  class Base
+    def alipay?(expected_version = nil)
+      Alipay.new(ua).match? && detect_version?(full_version, expected_version)
+    end
+  end
+
   class Alipay < Base
     MATCH_REGEX = /AlipayClient/i
     VERSION_REGEX = %r[(?:AlipayClient)/([\d.]+)]i
@@ -21,4 +27,6 @@ module Browser
       ua =~ MATCH_REGEX
     end
   end
+
+  matchers.unshift(Alipay)
 end
