+++ lib/browser/blackberry.rb
@@ -24,4 +24,6 @@ module Browser
       ua =~ MATCH_REGEX
     end
   end
+
+  matchers.unshift(BlackBerry)
 end
