+++ lib/browser/electron.rb
@@ -1,6 +1,13 @@
 # frozen_string_literal: true
 
 module Browser
+  class Base
+    # Detect if the browser is Electron.
+    def electron?(expected_version = nil)
+      Electron.new(ua).match? && detect_version?(full_version, expected_version)
+    end
+  end
+
   class Electron < Base
     VERSION_REGEX = %r[Electron/([\d.]+)]
     MATCH_REGEX = /Electron/
@@ -22,4 +29,6 @@ module Browser
       ua =~ MATCH_REGEX
     end
   end
+
+  matchers.unshift(Electron)
 end
