+++ lib/browser/facebook.rb
@@ -1,6 +1,14 @@
 # frozen_string_literal: true
 
 module Browser
+  class Base
+    # Detect if browser if Facebook.
+    def facebook?(expected_version = nil)
+      Facebook.new(ua).match? &&
+        detect_version?(full_version, expected_version)
+    end
+  end
+
   class Facebook < Base
     VERSION_REGEX = %r[FBAV/([\d.]+)]
     MATCH_REGEX = /FBAV/
@@ -21,4 +29,6 @@ module Browser
       ua =~ MATCH_REGEX
     end
   end
+
+  matchers.unshift(Facebook)
 end
