+++ lib/browser/meta.rb
@@ -4,11 +4,6 @@ require "browser/meta/base"
 require "browser/meta/generic_browser"
 require "browser/meta/id"
 require "browser/meta/ie"
-require "browser/meta/ios"
-require "browser/meta/mobile"
-require "browser/meta/platform"
-require "browser/meta/proxy"
 require "browser/meta/safari"
 require "browser/meta/webkit"
-require "browser/meta/tablet"
-require "browser/meta/device"
+require "browser/meta/proxy"
