+++ lib/browser/micro_messenger.rb
@@ -1,6 +1,15 @@
 # frozen_string_literal: true
 
 module Browser
+  class Base
+    # Detect if browser is MicroMessenger.
+    def micro_messenger?(expected_version = nil)
+      MicroMessenger.new(ua).match? &&
+        detect_version?(full_version, expected_version)
+    end
+    alias_method :wechat?, :micro_messenger?
+  end
+
   class MicroMessenger < Base
     VERSION_REGEX = %r[(?:MicroMessenger)/([\d.]+)]i
     MATCH_REGEX = /MicroMessenger/i
@@ -21,4 +30,6 @@ module Browser
       ua =~ MATCH_REGEX
     end
   end
+
+  matchers.unshift(MicroMessenger)
 end
