+++ lib/browser/nokia.rb
@@ -1,6 +1,13 @@
 # frozen_string_literal: true
 
 module Browser
+  class Base
+    # Detect if browser is Nokia S40 Ovi Browser.
+    def nokia?(expected_version = nil)
+      Nokia.new(ua).match? && detect_version?(full_version, expected_version)
+    end
+  end
+
   class Nokia < Base
     VERSION_REGEX = %r[S40OviBrowser/([\d.]+)]
     MATCH_REGEX = /S40OviBrowser/
@@ -21,4 +28,6 @@ module Browser
       ua =~ MATCH_REGEX
     end
   end
+
+  matchers.unshift(Nokia)
 end
