+++ lib/browser/otter.rb
@@ -1,6 +1,14 @@
 # frozen_string_literal: true
 
 module Browser
+  class Base
+    # Detect if browser is Otter.
+    def otter?(expected_version = nil)
+      Otter.new(ua).match? &&
+        detect_version?(full_version, expected_version)
+    end
+  end
+
   class Otter < Base
     VERSION_REGEX = %r[Otter/([\d.]+)]
     MATCH_REGEX = /Otter/
@@ -21,4 +29,6 @@ module Browser
       ua =~ MATCH_REGEX
     end
   end
+
+  matchers.unshift(Otter)
 end
