+++ lib/browser/phantom_js.rb
@@ -1,6 +1,14 @@
 # frozen_string_literal: true
 
 module Browser
+  class Base
+    # Detect if browser is PhantomJS
+    def phantom_js?(expected_version = nil)
+      PhantomJS.new(ua).match? &&
+        detect_version?(full_version, expected_version)
+    end
+  end
+
   class PhantomJS < Base
     VERSION_REGEX = %r[PhantomJS/([\d.]+)]
     MATCH_REGEX = /PhantomJS/
@@ -21,4 +29,6 @@ module Browser
       ua =~ MATCH_REGEX
     end
   end
+
+  matchers.unshift(PhantomJS)
 end
