+++ lib/browser/uc_browser.rb
@@ -1,6 +1,14 @@
 # frozen_string_literal: true
 
 module Browser
+  class Base
+    # Detect if browser is UCBrowser.
+    def uc_browser?(expected_version = nil)
+      UCBrowser.new(ua).match? &&
+        detect_version?(full_version, expected_version)
+    end
+  end
+
   class UCBrowser < Base
     VERSION_REGEX = %r[UCBrowser/([\d.]+)]
     MATCH_REGEX = /UCBrowser/
@@ -21,4 +29,6 @@ module Browser
       ua =~ MATCH_REGEX
     end
   end
+
+  matchers.unshift(UCBrowser)
 end
