+++ lib/browser/weibo.rb
@@ -1,6 +1,12 @@
 # frozen_string_literal: true
 
 module Browser
+  class Base
+    def weibo?(expected_version = nil)
+      Weibo.new(ua).match? && detect_version?(full_version, expected_version)
+    end
+  end
+
   class Weibo < Base
     MATCH_REGEX = /__weibo__/i
     VERSION_REGEX = %r[(?:__weibo__)([\d.]+)]i
@@ -21,4 +27,6 @@ module Browser
       ua =~ MATCH_REGEX
     end
   end
+
+  matchers.unshift(Weibo)
 end
