+++ test/test_helper.rb
@@ -13,10 +13,6 @@ require "yaml"
 
 module Minitest
   class Test
-    setup do
-      Browser::Bot.instance_variable_set("@detect_empty_ua", false)
-    end
-
     def assert_deprecated(message, file, line, &block)
       message = "#{message} (called from #{file}:#{line})"
       _out, err = capture_io(&block)
